Update daily top rankers to track cash prizes
