+++ backend/src/controllers/challengeController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import { db } from '../config/firebase.js';
-import { FieldValue } from 'firebase-admin/firestore';
+import { recordDailyWinnings } from '../utils/dailyRanking.js';
 
 interface ChallengeEntry {
   userId: string;
@@ -33,17 +33,6 @@ const checkTimeLimitAndUpdate = async (
   return entry
 }
 
-const recordDailyWin = async (userId: string) => {
-  const date = new Date().toLocaleDateString('en-CA', { timeZone: 'Asia/Kolkata' });
-  const rankingRef = db.collection('daily-rankings').doc(date);
-  await rankingRef.set(
-    {
-      [userId]: FieldValue.increment(1),
-      lastUpdated: new Date().toISOString(),
-    },
-    { merge: true },
-  );
-};
 
 export const createChallenge = async (req: Request, res: Response) => {
   try {
@@ -474,7 +463,7 @@ export const submitAnswer = async (req: Request, res: Response) => {
         );
       });
 
-      await recordDailyWin(userId);
+      await recordDailyWinnings(userId, reward);
     }
 
     let nextQuestion: any = null;
@@ -518,8 +507,8 @@ export const getDailyRankings = async (_req: Request, res: Response) => {
     const data = doc.data() as Record<string, number>;
     const entries = Object.entries(data)
       .filter(([key]) => key !== 'lastUpdated')
-      .map(([userId, wins]) => ({ userId, wins: Number(wins) }))
-      .sort((a, b) => b.wins - a.wins);
+      .map(([userId, amount]) => ({ userId, totalPrize: Number(amount) }))
+      .sort((a, b) => b.totalPrize - a.totalPrize);
     res.json(entries);
   } catch (error) {
     console.error('Error fetching daily rankings:', error);
